+++ modules/renter/contractor/contractmaintenance.go
@@ -905,9 +905,6 @@ func (c *Contractor) threadedContractMaintenance() {
 	// renew and how much extra funds to renew them with.
 	for _, contract := range c.staticContracts.ViewAll() {
 		c.log.Debugln("Examining a contract:", contract.HostPublicKey, contract.ID)
-		// Start by unregistering any alerts
-		c.staticAlerter.UnregisterAlert(modules.AlertIDContractMaintenanceRequired(contract.ID))
-
 		// Skip any host that does not match our whitelist/blacklist filter
 		// settings.
 		host, _, err := c.hdb.Host(contract.HostPublicKey)
@@ -1027,13 +1024,20 @@ func (c *Contractor) threadedContractMaintenance() {
 	}
 	c.log.Debugln("Remaining funds in allowance:", fundsRemaining)
 
-	// Register the AllowanceLowFunds alert if necessary.
+	// Register or unregister and alerts related to contract renewal or
+	// formation.
 	var registerLowFundsAlert bool
+	var renewErr error
 	defer func() {
 		if registerLowFundsAlert {
-			c.staticAlerter.RegisterAlert(modules.AlertIDRenterAllowanceLowFunds, AlertMSGAllowanceLowFunds, AlertCauseInsufficientFunds, modules.SeverityWarning)
+			c.staticAlerter.RegisterAlert(modules.AlertIDRenterAllowanceLowFunds, AlertMSGAllowanceLowFunds, AlertCauseInsufficientAllowanceFunds, modules.SeverityWarning)
 		} else {
 			c.staticAlerter.UnregisterAlert(modules.AlertIDRenterAllowanceLowFunds)
+		}
+		if renewErr != nil {
+			c.staticAlerter.RegisterAlert(modules.AlertIDRenterContractRenewalError, AlertMSGFailedContractRenewal, renewErr.Error(), modules.SeverityError)
+		} else {
+			c.staticAlerter.UnregisterAlert(modules.AlertIDRenterContractRenewalError)
 		}
 	}()
 	// Go through the contracts we've assembled for renewal. Any contracts that
@@ -1054,8 +1058,6 @@ func (c *Contractor) threadedContractMaintenance() {
 		if renewal.amount.Cmp(fundsRemaining) > 0 || c.staticDeps.Disrupt("LowFundsRenewal") {
 			c.log.Println("Skipping renewal because there are not enough funds remaining in the allowance", renewal.id, renewal.amount, fundsRemaining)
 			registerLowFundsAlert = true
-			// Register alert for this contract
-			c.staticAlerter.RegisterAlert(modules.AlertIDContractMaintenanceRequired(renewal.id), AlertMSGContractRenewalRequired(renewal.id), AlertCauseInsufficientFunds, modules.SeverityWarning)
 			continue
 		}
 
@@ -1065,8 +1067,7 @@ func (c *Contractor) threadedContractMaintenance() {
 		fundsSpent, err := c.managedRenewContract(renewal, currentPeriod, allowance, blockHeight, endHeight)
 		if err != nil {
 			c.log.Println("Error renewing a contract", renewal.id, err)
-			// Register alert for this contract
-			c.staticAlerter.RegisterAlert(modules.AlertIDContractMaintenanceRequired(renewal.id), AlertMSGContractRenewalRequired(renewal.id), AlertCauseFailedContractRenewal, modules.SeverityWarning)
+			renewErr = errors.Compose(renewErr, err)
 			continue
 		} else {
 			c.log.Println("Renewal completed without error")
@@ -1097,8 +1098,6 @@ func (c *Contractor) threadedContractMaintenance() {
 		if renewal.amount.Cmp(fundsRemaining) > 0 || c.staticDeps.Disrupt("LowFundsRefresh") {
 			c.log.Println("skipping refresh because there are not enough funds remaining in the allowance", renewal.amount, fundsRemaining)
 			registerLowFundsAlert = true
-			// Register alert for this contract
-			c.staticAlerter.RegisterAlert(modules.AlertIDContractMaintenanceRequired(renewal.id), AlertMSGContractRefreshRequired(renewal.id), AlertCauseInsufficientFunds, modules.SeverityWarning)
 			continue
 		}
 
@@ -1108,8 +1107,7 @@ func (c *Contractor) threadedContractMaintenance() {
 		fundsSpent, err := c.managedRenewContract(renewal, currentPeriod, allowance, blockHeight, endHeight)
 		if err != nil {
 			c.log.Println("Error refreshing a contract", renewal.id, err)
-			// Register alert for this contract
-			c.staticAlerter.RegisterAlert(modules.AlertIDContractMaintenanceRequired(renewal.id), AlertMSGContractRefreshRequired(renewal.id), AlertCauseFailedContractRenewal, modules.SeverityWarning)
+			renewErr = errors.Compose(renewErr, err)
 		}
 		fundsRemaining = fundsRemaining.Sub(fundsSpent)