+++ CHANGELOG.md
@@ -13,7 +13,7 @@ Latest:
 
 ### v1.4.2.1
 **Key Updates**
- - Added alerts for contracts that can't be renewed or refreshed
+ - Added alerts for when contracts can't be renewed or refreshed
 
 **Key Updates**
  - Wallet can generate an address before it finishes scanning the blockchain